+++ package.json
@@ -8,15 +8,16 @@
   "license": "MIT",
   "private": true,
   "dependencies": {
-    "@chakra-ui/core": "^0.8.0",
+    "@chakra-ui/react": "^1.6.12",
     "@emotion/core": "^10.0.35",
-    "@emotion/styled": "^10.0.27",
+    "@emotion/react": "^11",
+    "@emotion/styled": "^11",
     "emotion-theming": "^10.0.27",
+    "framer-motion": "^4",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-feather": "^2.0.8",
-    "react-router": "^6.0.0-alpha.3",
-    "react-router-dom": "^6.0.0-alpha.3",
+    "react-router-dom": "6",
     "react-scripts": "3.4.3",
     "swr": "^0.3.0",
     "timeago.js": "^4.0.2"
