+++ src/components/home.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { Flex, Box, Text, Stack, Link } from "@chakra-ui/core";
+import { Flex, Box, Text, Stack, Link } from "@chakra-ui/react";
 import { ArrowRight } from "react-feather";
 import { Link as BrowserLink } from "react-router-dom";
