+++ src/index.js
@@ -1,18 +1,19 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter as Router } from "react-router-dom";
-import { ThemeProvider, CSSReset } from "@chakra-ui/core";
+import { ChakraProvider } from "@chakra-ui/react"
 
-import App from "./components/app";
+import App from "containers/AppPage";
+
+const MOUNT_NODE = document.getElementById("root");
 
 ReactDOM.render(
   <React.StrictMode>
     <Router>
-      <ThemeProvider>
-        <CSSReset />
+      <ChakraProvider>
         <App />
-      </ThemeProvider>
+      </ChakraProvider>
     </Router>
   </React.StrictMode>,
-  document.getElementById("root")
+  MOUNT_NODE
 );
